requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
django-postgresql
psycopg2-binary